Profile page code split into more files
